Rename CRIMSON.MAIN to Crimson.Core
