+++ CRIMSON/Core/CrimsonCmdArguments.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using CommandLine;
 
-namespace CRIMSON.MAIN
+namespace Crimson.Core
 {
     internal class CrimsonCmdArguments
     {
